Created bookmarklet.js
This is a bookmarklet designed to work with a query such as DiscretePlot[Table[X~Binomial(j, 0.01), P(X>=b), {b, 1, 5}], {j, 1, 1000}] on Wolfram alpha. 
It allows you to approximate the x value of any point on the graph and the % at that x and y point. 
But it requires you to calibrate it first. 
You have to put your cursor in the top left part of the graph (0, 1.0), and press W. 
Then put your cursor in the bottom right part of the graph (in this case, (1000,0)) and press S. 
This calibrates it. 

However, if you are going from 1, to something other than 500 you will want to put that new number in the text field for the adjustment of the width. 
Then you click the button to adjust it, and now it should work.